requirements.txt
aioredis==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
black==22.3.0
certifi==2023.7.22
cffi==1.15.1
channels==3.0.5
channels-redis==3.4.0
charset-normalizer==2.1.1
click==8.1.7
constantly==15.1.0
cryptography==41.0.3
daphne==3.0.2
Django==4.0.7
django-extensions==3.2.3
django-ipware==5.0.0
gunicorn==21.2.0
h11==0.14.0
hiredis==2.2.3
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
msgpack==1.0.5
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==23.2.0
python-dotenv==1.0.0
quick-mailer==2022.2.22
requests==2.28.1
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.7.1
urllib3==1.26.16
uvicorn==0.23.2
zope.interface==6.0
